refactor: consolidate port parser tests into internal/ports and un-export parsers
Move fixture tests from internal/inventory into internal/ports, merge them
with the existing inline tests, and revert ParseLinuxOutput/ParseDarwinOutput
to unexported names. The only non-test callers were within internal/ports
itself, so the exported names were unnecessary.

- Move testdata/{ss-linux,lsof-macos}.txt to internal/ports/testdata/
- Add fixture-driven tests using findPortByNumber lookup instead of
  index-based ports[0]/ports[2] coupling
- Drop overlapping inline-string tests, keep edge-case coverage
- Delete internal/inventory/portparse_fixture_test.go
