+++ internal/ports/testdata/lsof-macos.txt
